refactor: divide function
