+++ src/functions/slack-webhook/mock.json
@@ -1,0 +1,26 @@
+{
+  "owner": "123456789012",
+  "logGroup": "CloudTrail",
+  "logStream": "123456789012_CloudTrail_us-east-1",
+  "subscriptionFilters": [
+    "Destination"
+  ],
+  "messageType": "DATA_MESSAGE",
+  "logEvents": [
+    {
+      "id": "31953106606966983378809025079804211143289615424298221568",
+      "timestamp": 1432826855000,
+      "message": "{\"eventVersion\":\"1.03\",\"userIdentity\":{\"type\":\"Root\"}"
+    },
+    {
+      "id": "31953106606966983378809025079804211143289615424298221569",
+      "timestamp": 1432826855000,
+      "message": "{\"eventVersion\":\"1.03\",\"userIdentity\":{\"type\":\"Root\"}"
+    },
+    {
+      "id": "31953106606966983378809025079804211143289615424298221570",
+      "timestamp": 1432826855000,
+      "message": "{\"eventVersion\":\"1.03\",\"userIdentity\":{\"type\":\"Root\"}"
+    }
+  ]
+}
